+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+## v0.1.18
+
+- Removed custom IOStream
+- Handle refreshing the environment after pip install
+- Dev dependencies updates
+
 ## v0.1.17
 
 - Updated ag2 to 0.5.3